+++ src/test/java/io/github/redouane59/twitter/nrt/ITwitterClientV2Test.java
@@ -588,13 +588,4 @@ public class ITwitterClientV2Test {
     assertNotNull(lists.getIncludes().getUsers().get(0).getCreatedAt());
   }
 
-  @Test
-  public void testLookUpAndGetSpaceBuyers() {
-    SpaceList result  = twitterClient.searchSpaces("hello", SpaceState.LIVE);
-    String    spaceId = result.getData().get(0).getId();
-    UserList  buyers  = twitterClient.getSpaceBuyers(spaceId);
-    assertNotNull(buyers.getData());
-    assertNotNull(buyers.getMeta());
-  }
-
 }
